+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "code-stamp" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+### Added
 
-- Initial release
+- Display additional information on hover, including file creation time.
+- Support commenting in files based on contributions from other extension languages.
+
+
+## [Version 1.0.0] - 2023-11-13